Respect default values provided to get() methods
The implementations of get() on Namespace and Experiment take a
defaultValue parameter which eventually filters down and is passed to an
Assignment.get() call. However, Assignment.get() does not take a second
argument and the defaultValue is summarily ignored.

These changes update Assignment.get() to take a second argument for
defaultValue and it maintains the same semantics as Interpreter.get():
if the named param is not found or is found but it is null or undefined,
get() will return defaultValue.