Use only built library code for tests
This removes cases where we would require() src code in the unit tests
and then try to intertwine those objects with built library code. We
should only be using built library code in the unit tests.

Effects of this commit:
* PlanOut.Ops.Core is no longer undefined
* Exposed the following additional modules on PlanOut:
	* PlanOut.Lib.Utils
	* PlanOut.Ops.Base
	* PlanOut.Ops.Utils
* Unit tests now pass on Windows machines

One trade-off here is that we now expose some previously internal
modules to clients of the library. In some cases, it seems like the
right thing to do. For instance, the unit tests demonstrate how to
register a custom operator, but without PlanOut.Ops.Base one cannot
replicate that in application code. In other cases, like
PlanOut.Lib.Utils, it seems like we might be exposing too much. I think
it may still be ok to do this, as anyone using this library will have
better alternatives for the internal functionality (lodash, underscore),
but I can understand if we want to find an alternative solution.