Allow for falsy WeightedChoice choices
If configured with a falsy choice, the WeightedChoice operator would
never select that choice. Instead, it would skip it and return the next
choice, thus essentially transferring all of the falsy choices's weight
to the next choice in line. This is especially dangerous because 0 is
conventionally designated as the "control" group. Any experiments
configured this way would unwittingly lose their control group.

This change modifies the implementation to be more closely aligned with
the Python reference implementation.

One caveat, when a WeightedChoice containing null or undefined choices
is applied to an Assignment, if those values are selected, they will be
considered "misses" and the defaultValue will be used instead, if
provided (or undefined if not).