Clean up exports (and imports)
For modules that export more than one item, we now use the
`export {...}` syntax. The corresponding imports are updated to the
`import * as <something> from ...`  syntax. This makes it more clear
and consistent that each module is exporting a collection of items
rather than a single item with a collection of properties.